% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_yule_coef}
\alias{compute_yule_coef}
\title{Compute Yule's Phi-coefficient}
\usage{
compute_yule_coef(z_star, z)
}
\arguments{
\item{z_star}{a true block membership}

\item{z}{an estimated block membership}
}
\value{
Real value of Yule's Phi-coefficient between the true and estimated block membership is returned.
}
\description{
Compute Yule's Phi-coefficient
}
\examples{
data(toyNet)
compute_yule_coef(z_star = toyNet\%v\% "block",
                  z = sample(c(1:4),size = 200,replace = TRUE))
                  
}
